waitpid working

+++ Userland/SampleCodeModule/include/processlib.h
@@ -21,21 +21,21 @@
 
     int64_t spawn_process(int64_t function_address, int argc, char ** argv, unsigned int priority, bool foreground);
 
-    bool kill(uint64_t pid);
+    bool kill(int64_t pid);
 
     uint64_t current_pid();
 
     void giveup_cpu();
 
-    bool blockp(uint64_t pid);
+    bool blockp(int64_t pid);
 
-    bool unblockp(uint64_t pid);
+    bool unblockp(int64_t pid);
 
-    bool nice(uint64_t pid, int prio);
+    bool nice(int64_t pid, int prio);
 
     void waitp();
 
-    void waitpid(unsigned int pid);
+    void waitpid(int64_t pid);
 
     ps_t * ps();
